chore(cmd): Remove IntVar and move to Int64Var

+++ internal/cmd/base/base.go
@@ -128,7 +128,7 @@ type Command struct {
 	FlagAuthMethodId      string
 	FlagHostCatalogId     string
 	FlagCredentialStoreId string
-	FlagVersion           int
+	FlagVersion           int64
 	FlagRecursive         bool
 	FlagFilter            string
 	FlagTags              map[string][]string
